styling recovery and multiple import

+++ server.js
@@ -10,6 +10,7 @@ const jsonfile = require("jsonfile")
 
 app.use(bodyParser.urlencoded({ extended: false }))
 app.use(bodyParser.json({ limit: "40MB" }))
+app.use(cors())
 
 var listener = app.listen(7080, function () {
   console.log("Your app is listening on port " + listener.address().port)
@@ -23,18 +24,129 @@ app.get("/getall", async function (req, res) {
   res.send(vals)
 })
 
+function isString(value) {
+  return typeof value === "string" || value instanceof String
+}
+
+function isObject(value) {
+  try {
+    JSON.parse(value)
+    return true
+  } catch (error) {
+    return true
+  }
+}
+
+function patternCheck(mock) {
+  const getTemplate = {
+    endpoint: "",
+    response: {},
+    method: "get",
+    key: "get",
+  }
+
+  let toBereturnedObj = {}
+
+  const postTemplate = {
+    endpoint: "test",
+    request: {},
+    response: {},
+    method: "get",
+    key: "gettest",
+  }
+  const getTemplateKeys = Object.keys(getTemplate)
+  const postTemplateKeys = Object.keys(postTemplate)
+  const incomingMockKeys = Object.keys(mock)
+
+  if (mock.method === "get") {
+    toBereturnedObj = getTemplate
+    if (incomingMockKeys.length < 4) {
+      return false
+    }
+    for (let index = 0; index < getTemplateKeys.length; index++) {
+      toBereturnedObj[getTemplateKeys[index]] = mock[getTemplateKeys[index]]
+      if (index === 1) {
+        if (!isObject(toBereturnedObj[getTemplateKeys[index]])) {
+          return false
+        }
+      } else {
+        if (!isString(toBereturnedObj[getTemplateKeys[index]])) {
+          return false
+        }
+      }
+    }
+    // keys are very important. Must be validated carefully
+    toBereturnedObj.key = toBereturnedObj.method + toBereturnedObj.endpoint
+
+    return toBereturnedObj
+  } else {
+    toBereturnedObj = postTemplate
+    if (incomingMockKeys.length < 5) {
+      return false
+    }
+    for (let index = 0; index < postTemplateKeys.length; index++) {
+      toBereturnedObj[postTemplateKeys[index]] = mock[postTemplateKeys[index]]
+
+      if (index === 1 || index === 2) {
+        if (!isObject(toBereturnedObj[postTemplateKeys[index]])) {
+          return false
+        }
+      } else {
+        if (!isString(toBereturnedObj[postTemplateKeys[index]])) {
+          return false
+        }
+      }
+    }
+    // keys are very important. Must be validated carefully
+    toBereturnedObj.key = toBereturnedObj.method + toBereturnedObj.endpoint
+    return toBereturnedObj
+  }
+}
+
 app.post("/upload", async function (req, res) {
   // console.log(requestsWeOwn);
- const val = req.body;
-  console.log(val);
- 
+  let responseMessage = "Json File is not valid."
+  let requestsWeOwn = await db.get("allRequests")
 
+  const apis = req.body.body.apis
+  if (apis.length < 1) {
+    responseMessage = "false"
+    res.send(responseMessage)
+  }
 
-  res.send("true")
+  if (requestsWeOwn !== null) {
+    for (let i = 0; i < apis.length; i++) {
+      let element = patternCheck(apis[i])
+      if (!element) {
+        continue
+      }
+      // array increases every time but it prevents saving same request template multiple times
+      let updateflag = false
+      for (let index = 0; index < requestsWeOwn.length; index++) {
+        if (requestsWeOwn[index].key === element.key) {
+          updateflag = true
+          requestsWeOwn[index] = element
+        }
+      }
+      if (!updateflag) {
+        requestsWeOwn.push(element)
+      }
+    }
+  } else {
+    requestsWeOwn = []
+    for (let j = 0; j < apis.length; j++) {
+      let element = patternCheck(apis[j])
+      if (!element) {
+        continue
+      }
+      requestsWeOwn.push(element)
+    }
+  }
+  await db.delete("recover")
+  await db.set("allRequests", requestsWeOwn)
+
+  res.send(true)
 })
-
-
-
 
 app.get("/exportall", async function (req, res) {
   const vals = await db.fetch("allRequests")
@@ -74,6 +186,7 @@ app.post("/savetemplate", async function (req, res) {
     requestsWeOwn.push(req.body.body)
   }
   try {
+    await db.delete("recover")
     await db.set("allRequests", requestsWeOwn)
   } catch (err) {
     console.log(err)
@@ -85,13 +198,30 @@ app.post("/savetemplate", async function (req, res) {
 
 app.get("/cascadeall", async function (req, res) {
   // ne var ne yok temizler
-  await db.delete("allRequests")
+  let recover = await db.fetch("allRequests")
+  if (recover === null) {
+    res.send(false)
+  } else {
+    await db.set("recover", recover)
+    await db.delete("allRequests")
+    res.send(true)
+  }
+})
 
-  res.send("True")
+app.get("/recover", async function (req, res) {
+  // returns to the backup after cascade operation
+  let recover = await db.fetch("recover")
+  if (recover === null) {
+    res.send(false)
+  } else {
+    await db.set("allRequests", recover)
+    await db.delete("recover")
+    res.send(true)
+  }
 })
 
 app.get("/mocktail/:endpoint", async function (req, res) {
-  // validasyon yazılacak
+  // validate ?
   let potentialResponse = { status: "404 endpoint does not exist." }
   let vals = await db.fetch("allRequests")
   for (let index = 0; index < vals.length; index++) {
@@ -156,4 +286,8 @@ app.get("/delete/:endpointkey", async function (req, res) {
   res.send(potentialResponse)
 })
 
-app.use(cors())
+// app.get("/del", async function (req, res) {
+//   // let potentialResponse = { status: "Could not delete." }
+
+//   res.send(fefe)
+// })
